fix: expose api config to enable using external clients and APIs (#1523)
